flaky_proxy: Evaluate the block passed to match on each request
This means that the built-in sequencing behavior will no longer work as before,
but this model seems to make more sense upon first usage, and is more flexible
that the previous one.

We should also be able to add some nice syntax for defining sequences as before
as a follow-up.

Bonus: the request and connection are now available as parameters to the block
passed to match, allowing you to easily construct responses based on the content
of the request.