RUBY-1242 Reset transactions at start of requests
For Rack-based web requests, we will force the transaction to be cleared out
as early as possible in our request pipeline. This leans on our middlewares,
all of which will attempt to reset if no one else has done in this request
(tracked by the env).

This makes it safe for us to use values from a completed transaction that are
still present in the thread local after the transaction is done.