RUBY-1253 Separate transaction event custom params
Previously we were merging the custom parameters directly into a single hash
for the transaction events. The collector now supports a format where we send
two hashes--one with the agent values, another with custom parameters--to avoid
naming collisions.

While it's still possible that colliding keys are sent, the collector is now
responsible for our policy on how to resolve those conflicts.