RUBY-1253 Safer exception handling on RUM loader
If we had an exception during generating the JS for either the loader or config
we would fail the users request. Not cool.

We now trap and log if anything goes badly wrong in those cases. This is done
in the JavascriptInstrumentor rather than the middleware to make sure that
clients using manual instrumentation also get the protection.

The primary worry is JSON serialization failures from our NewRelic.json_dump
call.