RUBY-1262 Better handling for unknown selectors
We whitelisted $cmd selectors in the prior release because we can't reliably
get the collection name from unordered hash selectors on older Rubies.

As a fallback, rather than recording as unknown, this commit sets
"NoCollection" for the collection, but does use the first selector's key as the
command name.

This doesn't have the nested-hash problems we had with automatically grabbing
the first _value_ as the collection name. Since Mongo itself "requires" hashes
to be ordered, that first key is the command even if its value isn't the
collection.