RUBY-1259 Wrappers for snapshotting VM values
In order to deal sanely across the various versions and implementations of Ruby
we support, this adds a layer that allows us to safely take snapshots of the VM
stats that we care about.

NewRelic::Agent::VM::Snapshot will have nil's for values that aren't supported
and represents the state of the VM at a given instant in time. It will be the
responsibility of the sampler to properly compare to prior snapshots and
calculate metrics based off these values.