RUBY-1263 Skip Sequel explain plans if single-threaded connection pool in use
If a single-threaded connection pool is in use with Sequel, we can't be sure
that using the connection pool from the harvest thread is going to be safe (e.g.
someone else may be using it concurrently), so the safest thing to do is just
skip collection of an explain plan in this situation. In theory we could capture
the options used to create the pool and create our own separate one just to run
the explain plan, but the additional risk and complexity seem not worth it.