Support setting multiverse temp directory
Setting ENV["MULTIVERSE_TMP"] will now allow overriding the default tmp
directory that our Mongo tests write to.

This is necessary for CI since we use that tmp directory to track which ports
we expect to use from the test servers we spin up. This defaults to the gem
folder, but we have multiple copies of the gem in CI which prevents them from
sharing information and introduces conflicts.