Start up Mongo if it isn't running
This is intended for local development, and uses the MongoServer to get things
running and shutdown in the before/after_suite.

Along with using MongoServer from the multiverse suite, this also adds a flag
to allow us to start the MongoServer without actually touching the mongo gem.
This is necessary since the multiverse master process that actually runs before
and after hooks for the suite doesn't have mongo--only the children do.

There is a potential for a race if starting Mongo isn't fast enough before the
tests run, but it hasn't proven to be an issue yet and we'll see if it is
before adding checks to the Envfile since it can't use mongo itself.