RUBY-1274 Record Datastore/all metric
As part of recording the background database time for attaching to analytic
events, we're also adding a full rollup metric for all database time--web and
background. This used to be ActiveRecord/all, but that only represents web time
since 3.6.3 to preserve the overview graphs.

There is no direct usage of this metric presently, it may help in solving some
consistency issues with the Database tab and app map in RPM.