RUBY-1273 Allow GUID on untraced transactions
If the transaction tracer was disabled, we didn't create a GUID since that was
being done on the sample, not the transaction itself.

With capturing GUID's on Dirac events, this prevented events from getting the
correlating GUID's, which would be suprising to a user.

Now we generate the GUID on creation of the transaction itself, and then read
that from the other pertinent locations.