Avoid race condition in pipe channel tests
We had a race in these tests that caused intermittent errors if the parent
didn't read off the pipe from the child quickly enough before the test shut
down the pipe channel thread.

While a little touchy to duplicate locally, I did get it to happen occasionally
and introducing even a small synthetic delay in the pipe reading was enough to
induce the same failures we saw before.

We now simply wait until the pipe has hit eof. Since we've already Process.wait
on the child closing out, this shouldn't need a timeout to make sure it gets
finished.