Fix spurious failures on JRuby
Occasionally the timings for really short local queries on JRuby would end up
as 0. For tests checking the explain plans, that would fail to generate an
explain since the time wasn't > the threshold.

Making the threshold negative, so we always take an explain, gets us where we
want to be for this particular test.