RUBY-1182 Coerce Rails.env to string
Certain versions of Rails return an ActiveSupport::StringInquirer rather than a
String. Additionally we've seen monkey-patches to JSON encoding that then work
poorly with that class.

Coercing the Rails.env to a String via to_s takes insulates us. If we find lots
of other cases where other types need coercing we may look for a generic
solution, but that doesn't seem necessary at this point.