Fix intermittent time Marshaling errors on JRuby
We occasionally will get failures in CI from JRuby incorrectly marshaling a
Time value in the StatsHash class. This might be related to some known issues
that are unresolved at this time. https://github.com/jruby/jruby/issues/957

To deal with this, we instead will start treating the StatsHash @started_at
ivar as a float since that's what we coerce it to for collector communication
anyway.

This is primarily exercised when marshaling from Resque child workers to the
parent, though, since we don't directly marshal the StatsHash objects to the
collector anymore.