RUBY-1247 More reliable hostname and started flag
When the agent is run under certain dispatchers (Unicorn, Resque, DJ) it will
defer startup during the gem initialization. In these cases we weren't setting
the @local_host variable on the agent. If under those conditions the agent was
started manually and NewRelic::Agent.after_fork was called, the connect would
fail because of the missing host name.

This showed up in a number of support cases where users had Unicorn with
another framework and were applying the option to restart the agent after_fork
in Unicorn without sufficiently stringent gating to keep the logic from being
hit by their background job startup.

We now lazily gather the host name, and also make sure to mark ourselves as
started when the after_fork method kicks up a new thread.