RUBY-1247 Consolidate actual thread startup steps
We have two similar but distinct paths in the startup--one for the actual main
start, another for the after_fork hook we call from certain dispatchers.

This commit moves the flagging of started so we'll be more easily able to
consolidate the two paths together. The merging of the two paths will be in a
separate commit.