RUBY-1166 Remove special Puma and Unicorn handling
This changes the setting for restarting threads in child processes to true.
Now the agent will try to start up in a new process on first transaction
complete if it hasn't started already.

The upshot of this is that we no longer need to have special instrumentation
for Puma to get worker mode functional (and daemonize works), and Unicorn no
longer requires us to have `preload_app true` or an after_fork hook.

The only downside of this change in functionality is that a site without
traffic under some conditions won't start up. Will see in testing if this is
problematic.