RUBY-1166 Don't mark start in Harvester#initialize
Original version of the Harvester class would implicitly take the PID of the
process it starts up in as the starting PID. This meant that deferring
processes would have to clear/reset the starting PID which was awkward.

Now we don't mark the starting PID until we actually get the thread running (or
close to it), so deferring processes don't have to do anything explicitly.