Make test:env version lookups less repetitive
Noticed running on 2.1.1 that we were incorrectly getting the 2.x versions of
Rails because test:env was checking for 2.1.0 explicitly. Logic now does
start_with? checks on versions to let us be a little more sweeping in our
assignment of appropriate ranges.

In addition this updates for 2.1.1 and corrects for Rails 4.1 showing up on
1.9.2.