RUBY-1166 Test after_fork's mark started directly
It's kind of a subtle thing that we call the Harvester#mark_started at the
start of after_fork when it does get triggered later by setup_and_start_agent
as well. The reasoning is that after_fork should only be called once we're in
a confirmed process where we're supposed to run. If we've gotten to after_fork
we shouldn't be trying thread restarts subsequently.

Given that subtle case, this commit adds tests to make that more explicit and
a comment on why we do this seemingly unnecessary thing at the start of
after_fork.