RUBY-1259 Fix GCProfiler tests, and make thread-safe
The previous implementation of GCProfiler (which we use to measure in-band GC
time) was not thread-safe. It stored the GC time and count snapshots in a global
singleton that could be overwritten in the middle of a transaction by another
thread.

We now instead store these snapshots on the Transaction itself, similar to
what's done for calculating CPU burn in the middle of a transaction. This allows
multiple concurrent transactions to all get correct GC timings.