Remove flaky testing of test control
This test is relying on config state that isn't actually modified by the tests
or the test control in question--it was working largely based on luck.

It also just mimics the exact conditionals in the test control and isn't really
a meaningful test of production functionality, so I'm just removing it.