RUBY-1276 Replace (or add) ServerSource
Originally the connect method was calling to simply apply_config with any newly
received ServerSource. If the agent restarted due to server-side config changes
this resulted in multiple ServerSources piling up.

We now replace any existing ServerSource when adding a new one (or simply add
it if it's the first instance).