RUBY-1259 Change the way we record VM gauge-type metrics
The heap_live, heap_free and thread count metrics will all be presented as
average values per instance, but the collector will aggregate all of the
reported values from each host into a single bucket. This means that we need a
convenient way of knowing how many instances contributed to a given timeslice.

Storing the measured gauge value (e.g. thread count or number of live objects)
in the total_call_time field of the metric, and a static value of 1 in the
call_count field will allow us to do this. In order to calculate an average
value per instance, we'll be able to just divide total_call_time by call_count
(since the collector will sum the 1's from each reporting instance up for us).

Note that we'll also need to change the metric parsers in rpm_site in order for
this to work correctly.