Fix for load ordering in Rails multiverse tests
Depending on the exact file order that things get required in (which varies per
system), we were getting some failures in CI on the Rails multiverse tests for
4.1.0.rc2.

These are a problem of wrong require orders that happened to work. This is now
centralized in `app.rb` for the Rails suite so it's always applied right (and
also pulls in `agent_helper` consistently for us).