RUBY-1188 Split GC/cumulative between web/other
Similar to past issues with database metrics, having large GC times in
background processes can result in swamping the web overview (i.e. missing or
too-small Ruby time).

This commit separates the metrics we record for GC between web and other
transactions. We discontinue recording GC/cumulative, and the overviews will
be updated to correctly display the new GC metrics.

The meaning of the GC data gathered is not changed by this, just the bucketing
of the values.