RUBY-1188 Revise call_count on new GC metrics
Since we're transitioning to a new set of GC metrics, we're taking the chance
to record them a little differently.

Prior versions tried to record the GC counts into the call_count field of the
metric. This resulted in us ignoring some GC time (for minor runs and sweeps)
that didn't increment the overall count. It also was the last remaining client
for the record_multiple_data_points method on Stats, which was a weird beast.

Now the elapsed GC time is always added into the total_time, and call_count
will be an (unused) proxy for the number of transactions we've actually
profiled GC against.