RUBY-1284 No nested transactions for Sinatra middlewares
If a Sinatra app is being used as a middleware (ala `use TheApp`), our
instrumentation was wrongly generating multiple nested transactions for each
of the Sinatra middlewares the request passed through.

Instead, we want to only start a transaction at the outermost point, and rely
on the already existing machinery to rename the transaction with the right
route once it's found.