Sidekiq 3.0.0 update
This updates multiverse to run with sidekiq 3.0.0 which was just released.

It also makes a small signature change because our tests use the middleware in
the client chain, which had a breaking change in the 3.0.0 release (see
https://github.com/mperham/sidekiq/commit/83aea0690e0bf1405661c1259701211881f702ea
for relevant details.)

This does not impact the server middleware we actually install when
instrumenting, but the change is compatible with either format (plus any other
parameter additions in the future).