RUBY-1284 Notice errors on nested Sinatra transactions
Our change to not start new nested transactions from Sinatra routes had a
subtle impact on noticing errors. If the outer transaction wasn't started by a
Sinatra app itself, but the inner route that matches is Sinatra-based, we miss
out on the errors.

This happens when we include New Relic's Rack instrumentation in a
Sinatra::Base derived class, since the outer `call` starts a transaction, so by
the time Sinatra is dispatching we've already got it running, but without the
trap to check env['sinatra.error'] afterward.