Provide better debug mode for multiverse
This mode can be invoked simply passing debug to the rake command, and will
run all the tests in the foreground where debuggers can interact with it.

Debug mode also will now include pry as well as the debugger gem.

Usage: rake test:multiverse[suite,debug]

Note that when in foreground mode, the output of failures at the end of the
entire run doesn't happen anymore (since we can't capture the output and still
let the debugger interact with you). Shouldn't be a big deal though.