Move trace_execution_scoped_header/footer into Transaction
This moves the responsibility for calling trace_execution_scoped_header and
_footer into the Transaction itself. Since a primary purpose of a transaction
is managing that TT node stack and recording metrics about itself, this makes
good sense and was duplicated in callers to Transaction.start/stop.

Changes to the tests deal with the fact that our change introduces a new layer
of TT nodes to users of the in_transaction helper, and don't represent an
actual change in the real data recorded.