RUBY-1292 Removed :force option from trace_execution_scoped.
This was also an option to add_method_tracer and perform_action_with_newrelic_trace.
It's unclear just how this option was intended to interact with disabled tracing,
specifically when an entire transaction was wrapped in a disable_all_tracing
block.

We don't know of any usages of this option, and believe that there are better
ways for users to gain this functionality.

The option is now marked as deprecated, and will emit a warning.