Fix a race condition in AgentThreadTest that led to hangs on 1.8.x, rbx
There was a race condition in the test_bucket_thread_as_request and
test_bucket_thread_as_background tests in AgentThreadTest that was leading to
hangs when running those tests under Ruby 1.8 and Rubinius.

The goal of these tests is to start a background thread, wait for it to reach
some state, attempt to identify what kind of thread it is, assert that the
identified type was correct, and then unblock and join the background thread.

This was being done with a sleep/run combo in order to hold the background
thread in a certain state until the foreground thread could try to bucket it.
Unfortunately, there was a race that could cause the run call to happen before
the sleep, leading to a deadlock. This was highly reproducible on Ruby 1.8.x and
Rubinius, though not on MRI 1.9+.

These tests now use two separate queues to signal from the main -> background
thread and vice-versa, which appears to resolve the race.