Config management cleanup.
The various configs were loaded and stored in a fragile, haphazard way. For
example, load order partially determined predence. We would see intermittant
test failures that we'd track down to config issues, like having multiple
ServerSource configs in the stack (which should not be possible).

Now, each of the 5 config types is stored in its own slot. It is not possible
to have more than one of any config type, and precedence is determined strictly
by config type (i.e. load order is irrelevant).