Remove the :scoped_metric_only option to trace_execution_scoped
This option is technically part of the public API, but it's listed under
'uncommon options', and I don't believe that there exist any legitimate use
cases for it. This change will effectively make passing the :scoped_metric_only
option into a no-op.

Consulting with other agent teams (PHP, Java, and .NET), they all record both
scoped and unscoped metrics for custom tracers. This is particularly important
because only unscoped metrics can currently be displayed on custom dashboards.

This doesn't have any direct performance impact, but it will make a future
refactoring that I have in the pipeline to metric recording easier to do.