Remove :deduct_call_time_from_parent option to trace_execution_scoped
This option is technically part of the public API, but listed as an 'uncommon'
option, and I don't believe it has any legitimate use cases. This change will
make passing it into a no-op.

The intent of this option is to allow you to control whether a traced segment
B invoked from within another traced segment A would subtract B's duration from
A's total_exclusive_time.

It turns out that we pretty much always want this. There have in fact been bugs
in the past where we've failed to set this option, and ended up with
double-counting.

This change itself has no significant performance impact, but it will make other
performance-related refactors in the pipeline easier.