RUBY-1292 Adding a TransactionMetrics class
This class will be used to store recorded metrics during the course of
a transaction, before they are merged into the global StatsHash.

The advantage of this class over using another StatsHash instance is that we
won't have to allocate MetricSpec instances to use as keys, since we just keep
the scoped and unscoped metrics in separate transactions.