Stop subclassing ::Thread for AgentThread
The only purpose behind AgentThread subclassing Thread was so we could
get the thread local labelling on startup of the thread. Doesn't require
us overriding the class to make that happen.

This changes over to just having a factory method (create) on the
AgentThread class that does the thread creation. It also adds wrapping
around the passed block to log errors and exits for us automatically,
which will be nice for things like the pipe and backtrace service which
weren't always doing that consistently.

Lastly, this makes the ThreadedTestCase slightly more sane, as
AgentThread now has a dedicated way to swap the backing thread class it
will create from, so no more constant shuffling.