Stop testing Sidekiq inline, run internal server
There are signficiant differences between Sidekiq's inline mode and how
things actuall work in practice. In particular server middlewares aren't
executed, which is a problem for our instrumentation (which installs
itself as a server middleware).

We've had issues in the past with process coordination, so instead this
commit changes over to using the Sidekiq::CLI class directly to run an
in-process server. This allows us to reach between tests and the jobs to
signal when we're actually finished rather than using kludgy timeouts.

Additionally this breaks out the test runs to use their own separate
queues, since running in parallel breaks badly when all jobs are writing
to the same shared queue. I wonder if this might have been the source of
our instability in earlier revs of the tests. A question for the ages...