RUBY-1309 Missing logging with manual start
If the agent is loaded but disabled by yml/env config at that point, a
subsequent call to NewRelic::Agent.manual_start would not replace the
logger instance. This yields a functioning, but log-silent agent which
would is massively confusing and difficult to debug.

Now if the initial disabled load (which results in a NullLogger) is
noticed, it's treated like the other startup loggers that will be
replaced with a real logging instance.