RUBY-340 Fixing flaky tests
We were getting intermittent failures because of configs getting changed
out from under the ErrorCollector instance, so it was missing getting
callbacks when the ignored errors key changed.

Also altered the protocol with Rails 4 reporting errors, so test needs
to notice the error to get the exception considered for apdex.