Try to use Addressable for Typhoeus URI parsing
Per https://github.com/newrelic/rpm/pull/150 Typhoeus will allow for
some URI's that aren't supported by the standard library URI parsing.
These can be read with Addressable, so if it's available, favor running
the URL's through that instead of std-lib ::URI.

This also introduces a conditional test in the HTTP client test cases.
This test can be enabled for a specific multiverse environment by
including the Addressable gem in the Envfile.