Stringify symbols before serializing, avoid ActiveSupport::JSON screwing with us
On some combinations of the JSON gem and ActiveSupport, attempting to JSON.dump
an object graph containing a Symbol will cause an error when the native code in
the JSON gem attempts to call out to a to_json method defined by ActiveSupport
that has a different signature than what the JSON gem code is expecting.

Serializing via the JSON gem in this situation is fine as long as the object
graph consists solely of basic types with directa anlogues in JSON (Array, Hash,
Fixnum, Float, NilClass, TrueClass, FalseClass, and String). The trouble arises
when the object graph contains some object not of one of these classes, and the
C code in the JSON gem tries to call out the to bogus to_json method that's been
added by ActiveSupport.

So, when we encounter a Symbol in the object graph to be serialized, we'll now
convert it to a String before calling into the JSON gem.