Clean up ownership of the Configuration::Manager
The Configuration::Manager is now sorted on the Agent, rather than in a class
variable on the Configuration::Instance module.

This maintains the fact that the manager is only instantiated once, while
eliminating the class variable access needed in order to get at the config.

This turns out to be a non-trivial perf win - it gets us an approximately 3.5%
win on the middleware performance tests, since we access the config so
frequently from various parts of the agent.