Start out with the metrics we actually want for middlewares
Instead of starting out middleware tracing segments with
a 'Controller/Middleware/Rack' prefix, and then renaming them when we hit
a better transaction name via nested_transaction_name, start with the metrics
that we normally will want for middleware (prefix of 'Middleware/Rack'), and
promote the outer-most one at the end of the transaction by adding
a 'Controller/' prefix if we haven't found anything better by that time.

The main motivation here is performance - this change is a 5% perf win on the
middleware tracing test.