Use string key in Rack env to indicate request was captured, rather than symbol
This fixes an issue on Ruby 1.8 where the CrossAppMonitor fell over trying to
upcase a Symbol that was used as a key in the Rack env hash.

Also, the key name is now 'newrelic.captured_request' (with a dot) to bring us
into compliance with the key naming requirements in the Rack spec.