Don't need to build request in ActionControllerSubscriber any more
Previously, we were relying on the before_call event, emitted by the AgentHooks
middleware in order to build a Rack::Request object from which to extract
headers, URI, etc. on Rails 4.

Now that we force the transaction to be started at the outer-most agent
middleware, this is no longer necessary - the AgentHooks middleware will create
the request for us and save it on the TransactionState.