Remove recording of Mongrel queue length metric
This commit stops us from recording the Mongrel/Queue Length metric. This is
a perf win of ~2% on the basic middleware instrumentation test.

Mongrel was last updated over 4 years ago, and we have limited ability to even
test against it anymore. In addition, the Mongrel/Queue Length metric is not
actually surfaced through the UI, so the only way anyone would actually see it
would be via a custom dashboard. Lastly, most customers who are still running
mongrel are on very old agent versions, so they likely won't be affected by the
removal of this metric anyway.