Don't store request parameters on slow SQL traces, since they aren't transmitted
The existing code went through some effort to attach request parameters that
were captured if capture_params was on to slow SQL traces, and made it seem like
the intent was eventually for them to be serialized and sent to the collector.
It turns out there's no logic in rpm_site to even show these if we did capture
them, so the existing code to keep request params on SQL traces is misleading at
best, and dangerous at worst.