Test capture params with middleware transactions
Previous tests for capture params were relying on a Rails controller
action to provide the main transaction. These tests add a check that
capturing the parameters works right when a middleware returns early so
we never hit the controller action at all.

The exact behavior differs a little between 2 and later versions because
of the NewRelic::Rack::ErrorCollector, but filtering is preserved where
params are captured, and all settings are obeyed correctly.