Extract common middleware tracing logic into MiddlwareTracing module
Previously, we had similar code in both the AgentMiddleware module and the
MiddlwareProxy class.

AgentMiddleware was used to trace our own middlewares (which need to establish
a transaction around themselves in order to work correctly) and MiddlewareProxy
was used to wrap 3rd-party middlewares (which wrapping could be disabled by the
disable_middleware_instrumentation setting).

Now, both AgentMiddleware and MiddlwareProxy rely on the MiddlewareTracing
module in order to host common functionality.

This has the added benefit of our 1st-party middlewares not using
ControllerInstrumentation, which is a ~5% perf win on the middleware perf test.