Handle failures during ignore_error_filter calls
Clients can custom error filtering behavior by passing a block that
decides whether to filter every individual exception we try to notice or
not. If an error occurred during evaluating that user-defined filter,
the current transaction was failing entirely.

Now instead we will treat a failure during the filter as implicitly
allowing the error and log verbosely about the problem.

This commit also adds a with_ignore_error_filter helper to agent_helpers
so tests can safely set the global filter and have it reverted properly
when the test is done (to prevent a failed test using setting a filter
from causing other tests downstream from failing too because of the
left-behind filter).