Don't call params on ControllerInstrumentation host unless no block given
Previously, ControllerInstrumentation would avoid calling the params method on
the host class to extract request parameters unless
perform_action_with_newrelic_trace was called without a block (this generally
only happens in a Rails 2 context).

We changed this recently, and it ended up breaking some tests in rpm_site where
we were including ControllerInstrumentation into a Grape application, which
responded to #params, but crashed when that method was called.

This change gets us back to the previous behavior of only calling params when
perform_action_with_newrelic_trace is called without a block, and adds tests
asserting that this remains the behavior going forward.