Restore ability to set segment_id on TransactionSample::Segment init
It turns out that rpm_site expects to be able to do this from tests, and is
relying on our TransactionSample and TransactionSample::Segment classes. Now
doesn't feel like the right time to attempt to disentangle that, since it
appears to go pretty deep in rpm_site, so this commit re-adds the ability to set
the segment_id on a TransactionSample::Segment from the initializer.