Don't use keys.include? when checking for existence of exception in hash
When multiple exceptions that are noticed by the agent occur during a single
transaction, we store them in a hash keyed by Exception (to avoid noticing the
same exception multiple times). In order to determine whether a given exception
e was already in the hash, we were using @exceptions.keys.include?(e). If it did
exist, we would attempt to merge in the options associated with the e with the
existing entry.

However, include? uses == in order to check equality, whereas Hash#[] uses eql?
This means it is possible for the include? check to return true, but the lookup
to return nil, resulting in us calling merge! against nil, if two distinct
exceptions that == each other are raised during the same transaction.

This was breaking the application_migrator_test in rpm_site.