Synchronize Sidekiq worker tests to transaction
We've had flakiness on the Sidekiq tests since we re-introduced actually
running them with the Sidekiq processor. Turns out the signaling at the
end of the job's perform method wasn't what we actually needed, since we
want to wait for the full transaction to complete. In rare cases we'd
signal at the job perform's end, but the transaction wouldn't be done
until after we'd already trigger transmitting data in the driving test.

The TestWorker now subscribes to the transaction_finished event and
signals when we reach the desired count there instead.

This commit also adds synchronization around the shared @jobs variable.