Restore block-based messages for StartupLogger
Noticed in the course of testing that I wasn't getting the debug log
output for the state of the configuration in the earliest stages, but
was getting it for the ServerSource.

Turns out that this broke when we made the config logging block-based,
since the StartupLogger (which runs before we have the full AgentLogger)
didn't know what to do with the blocks and ended up with only empty
messages to log.