RUBY-1318 Outermost middleware should always record Middleware/all metric
The previous fix for this issue only worked if the inner-most application was
a Rack app (not a Rails or Sinatra). This is because we were consulting the
frozen_name (that is, the name of the transaction as a whole) when determining
whether to record the Middleware/all summary metric for the outermost Rack
middleware. The frozen_name for a request that hit Rails or Sinatra would be
named after the Rails controller action or Sinatra route that was hit, and thus
it wouldn't start with with the middleware prefix we were looking for.