Change to Sidekiq synchronization in tests
Rather than only firing off a signal when we're done, instead Queue up
events for each finish and let the test itself wait on the proper number
with a long fail-safe timeout.

This should prove more stable in CI where the performance is much slower
than on a local system and we were hitting the condition variable
timeout relatively often.