RUBY-1323 Fix for exclusive time tracking
When removing the deduct_call_time_from_parent option from the top-level
transaction options, we missed that this value was being defaulted
based off whether a TT node recorded metrics or not. This meant that a
TT node that didn't record metrics was still getting removed from the
exclusive time of it's parent nodes, resulting in a skew to the
exclusive times in TT segments.

We don't want the full option back, but tracing does need to respect
that behavior around how we calculate the child times.