RUBY-1319 Default record_sql correctly
Per the spec, we need to allow a setting of 'obfuscated' for record_sql to
stay, but change any setting of 'raw' to 'obfuscated' instead. 'off' will
be left untouched.

Note that because of the order of config loading, this only takes into
account the local settings (yml + env) which seems like reasonable
behavior to me.