Use JSONWrapper.normalize instead of recreating it
We already had a standard method for normalizing JSON objects, so we
should just use that instead of creating our own "test" version.

JSONWrapper.normalize required two changes to work for this purpose.
First off, it wasn't converting symbol keys to strings, which is
problematic.

Second, it assumed if we got called through the normalization path we
were in 1.9 and could safely touch String#encoding, which isn't the case
now (at least in tests). To fix this, the string normalization method
now does the supports_normalization? check before proceeding to fiddle
with the string encodings.