RUBY-1331 Don't auto-inject RUM with ActionController::Live
Our auto-injection of RUM instrumentation causes issues when mixed with
ActionController::Live. In particular the way we close out the Rack
response is poorly timed, and can crash some apps while in others the
results would just be potentially very wrong.

Instead, if we note that ActionController::Live is included in a
controller action when we hit the instrumentation point, we will just
punt on doing the injection entirely. If a client wants RUM in a "live"
action, they'll need to manually instrument it.

Note that originally I would have liked to put this logic in the
ActionControllerSubscriber and just call Transaction#ignore_enduser!,
but because of how ActionController::Live spins off the action in
another thread that approach is totally unworkably non-deterministic.