RUBY-1327 Safer notice_error in transaction errors
In certain failure scenarios where we don't start up a transaction
properly, we could end up in a rescue block and try to notice_error on
the nil transaction. This would fail a request, so we shouldn't be doing
that.

NewRelic::Agent.notice_error takes the current transaction (or the
absence of one) into account appropriately already, so call that instead
in these rescues to stay safe.