RUBY-1333 Warn on 1.8.7 if no JSON found
Although JSON support isn't available in the standard library, upcoming
versions of the Ruby agent will be requiring the gem so we can stop Ruby
marshaling data across the wire.

This adds a deprecation warning so we can get the longest possible
lead-time on letting people know about the change. There will be lots of
other steps and communication before using the json gem is fully
required, but this is the first one.