RUBY-1325 Don't work so hard in the lock
When the transaction sampler is storing samples, we only really need to
synchronize around the actual writing of the ivar and storing the sample
in the buffers.

Given that trip and path calculations are actually kind of expensive
computationally, it's worth pulling those out of the synchronize to
minimize time locking.