Exclude commented YML lines from ERB processing
Because we run the newrelic.yml through ERB processing, if a user was to
comment out a line at the YML level that contains bad Ruby code in an
ERB template block, the file would fail to load. An example:

development:
  # <%= raise "O_o" %>

This commit now replaces any commented out line in the YML (whitespace
followed by #) with just # so ERB doesn't get run for the ignored lines.