RUBY-1334 Multiverse testing for labels
This commit adds multiverse testing for checking the actual transmitted
connect data for labels.

Getting this to work also took a bit of revision on DottedHash because
of inconsistencies in its treatment across the various source types.
YamlSource specifically would include the original keys pointing to
hashes, while other DottedHash subclasses (including those created by
the with_config test helper) did not.

We want those subhashes preserved most of the time across the dotted
flattening, the one exception being when we serialize the settings to
the collector, so we still support a mode of creation that doesn't
preserve the non-dotted keys.