RUBY-1334 Tighter validation on hashes
If we're taking the labels config directly from the newrelic.yml source,
it's possible for us to get handed things in the hash that aren't
strings and potentially are more deeply nested structures (arrays and
hashes). If that happens and those types manage to get past our
truncation (i.e. implements size), they could slip through and get sent
to the collector which would bomb out the connect with a RuntimeError,
leading to no data and continual retries on the connect from the agent
side.

This commit tightens up the validation and requires that our keys and
values be strings, no exceptions.