Stop calling super twice in ServerSource
Noticed that we were super'ing twice in ServerSource, which is just
weird.

Turns out this is pretty intentional, because it would:

1) Put the values into the hash itself (and later things like
   ungated_value were just talking to self[key] to look up keys)
2) Symbolize the keys so later code doing lookups could consistently
   find the desired values.

This changeset stops that, so we just build up the original hash
variable to contain what we want and super once at the end. A bit more
complicated, but keeps my eyes from crossing at the double-super calls.