RUBY-778 Lookup ignore filters if not there
If a user calls NewRelic::Agent.ignore_error_filter before the agent has
actually started (and hence instantiated the ErrorCollector), we would
miss out on assigning the filter properly. If you happened to call
ignore_error_filter after the agent was running you would get the
filtering.

We were already patching the incoming filter into a method on the
ErrorCollector class, so this commit just makes sure that we find it
there on the new ErrorCollector instance if we don't have one already.